updated design and split into components
