+++ code/src/index.css
@@ -1,5 +1,5 @@
 body {
-  background-color: lightyellow;
+  background-color: white;
   margin: 20px;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
     "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
@@ -14,7 +14,7 @@ code {
 }
 
 h1 {
-  color: plum;
+  color: #ffadad;
   font-size: 22px;
   display: block;
   width: fit-content;
@@ -23,13 +23,14 @@ h1 {
 }
 
 p {
-  margin: 10px;
-  color: plum;
+  margin: 10px auto;
+  max-width: 80%;
+  color: #000;
   font-size: 14px;
   font-weight: 500;
 }
 
-.container {
+.thoughtsContainer {
   display: flex;
   flex-direction: column;
 }
@@ -40,55 +41,69 @@ p {
   justify-content: center;
   align-items: center;
   width: 280px;
-  height: 150px;
-  border: 1px solid plum;
-  box-shadow: 5px 5px 2px plum;
-  margin: 5px auto;
+  height: fit-content;
+  border: 1px solid black;
+  box-shadow: 4px 6px black;
+  margin: 12px auto;
 }
 
 .newThoughtWrapper {
-  background-color: plum;
+  background-color: #f2f0f0;
   color: lightyellow;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
   width: 280px;
-  height: 150px;
-  margin: 0 auto 20px;
+  height: fit-content;
+  margin: 30px auto 20px;
+  border: 1px solid #000;
+  box-shadow: 5px 5px black;
 }
 
 .newThoughtWrapper label {
-  font-size: 18px;
+  font-size: 14px;
   font-weight: 500;
-  text-transform: uppercase;
+  color: #000;
+  margin: 10px auto 15px;
 }
 
-input {
-  width: 100px;
-  height: 20px;
+textarea {
+  width: 200px;
+  height: 50px;
   margin: 10px auto 0;
   font-size: 14px;
-  border: 1px solid purple;
+  border: 1px solid #fff;
   border-radius: 5px;
+  cursor: text;
+  caret-color: #ffadad;
 }
 
 .hearts {
-  font-size: 14px;
-  background-color: plum;
+  display: inline-block;
+  font-size: 12px;
+  background-color: #ffadad;
   border: none;
-  padding: 8px;
+  padding: 10px;
   border-radius: 50%;
+  cursor: cell;
+}
+
+.numberOfLikes {
+  display: inline-block;
+  font-size: 10px;
+  font-weight: 400;
 }
 
 .sendThought {
-  font-size: 14px;
-  background-color: lightyellow;
-  color: purple;
+  font-size: 12px;
+  background-color: #ffadad;
+  color: #000;
   border: none;
-  padding: 8px;
-  border-radius: 5px;
-  margin: 15px auto 0;
+  padding: 10px;
+  border-radius: 25px;
+  margin: 15px auto;
+  cursor: cell;
 }
 
 .date {
@@ -105,14 +120,20 @@ input {
     margin: 50px auto;
   }
   .container {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 10px;
+    width: 500px;
+    align-items: center;
   }
 
   .newThoughtWrapper {
-    width: 600px;
+    width: 500px;
     height: 200px;
-    margin: 20px auto 40px;
+    margin: auto;
+    align-items: center;
+  }
+
+  .thoughtWrapper {
+    width: 500px;
+    height: 200px;
+    margin: 30px auto 10px;
   }
 }
